### 2016-03-11 #28 @jotegui
rgeospatialquality
* 1. What does this package do? (explain in 50 words or less)
It allows users to assess the geospatial quality of biodiversity data by providing native access to the methods of the Geospatial Quality API (http://bit.ly/gqapi_doc_gh)
* 2. Paste the full DESCRIPTION file inside a code block below.

```
Package: rgeospatialquality
Title: R Wrapper for the Geospatial Data Quality REST API
Version: 0.2.2
Authors@R: person("Javier", "Otegui", email="javier.otegui@gmail.com",
    role=c("aut", "cre"))
Description: This package provides native wrappers for the functions available
    via de spatial quality REST API. More information on the API can be found
    here: http://bit.ly/bioinformatics_btw057.
License: file LICENSE
    jsonlite (>= 0.9.19),
    rgbif (>= 0.9.2)
Suggests: knitr,
    roxygen2,
```

* 3. URL for the package (the development repository, not a stylized html page)
https://github.com/jotegui/rgeospatialquality
* 4. What data source(s) does it work with (if applicable)?
None
* 5. Who is the target audience?
Biodiversity community: collection managers, researchers
* 6. Are there other R packages that accomplish the same thing? If so, what is different about yours?
There are no other packages, as far as I know
* 7. Check the box next to each policy below, confirming that you agree. These are mandatory.
    * [X] This package does not violate the Terms of Service of any service it interacts with.
    * [X] The repository has continuous integration with Travis CI and/or another service
    * [X] The package contains a vignette
    * [X] The package contains a reasonably complete README with `devtools` install instructions
    * [X] The package contains unit tests
    * [X] The package only exports functions to the NAMESPACE that are intended for end users
* 8. Do you agree to follow the [rOpenSci packaging guidelines](https://github.com/ropensci/packaging_guide)? These aren't mandatory, but we strongly suggest you follow them. If you disagree with anything, please explain.
    * [X] Are there any package dependencies not on CRAN?  No
    * [X] Do you intend for this package to go on CRAN?  Yes
    * [X] Does the package have a CRAN accepted license?    Yes
    * [X] Did `devtools::check()` produce any errors or warnings? If so paste them below.  No error or warning
* 9. Please add explanations below for any exceptions to the above:
* 10. If this is a resubmission following rejection, please explain the change in circumstances.
