### 2016-03-24 #28 @sckott
Nice package! Overall, this is great! Some suggestions for improvements detailed below.

## DESCRIPTION file

- `Title` field: you have `R Wrapper for the Geospatial Data Quality REST API`. Just change to `Wrapper for the Geospatial Data Quality REST API`, as they sometimes complain about the use of `R` in the title and description, since they say it's obvious it's for R since its on CRAN.
- `Description` field: you have `This package provides native wrappers for the functions available via de spatial quality REST API. More information on the API can be found here: http://bit.ly/bioinformatics_btw057.`. Change `de` to `the`

## Installation

Works on all platforms, windows and linux installs pass.

## travis file

I suggest removing

```
r_github_packages:
  - jimhester/covr
```

and put `covr` in `Suggests` in `DESCRIPTION` file

## license

Any reason you went with GPL? We usually default to MIT unless good reason to do something else. Just curious :)

also, docs https://cran.r-project.org/doc/manuals/r-release/R-exts.html#Licensing i think suggest that what you have (`file LICENSE`) isn't quite valid

## Imports

* You have 3 packges listed as imports, but you don't have roxygen import statements in your package.
The most obvious way to see this is that your imports aren't listed in your `NAMESPACE` file. Use `@importFrom pkg fxn1 fxn2 ...` to import just the functions you need, then when you run `document()` or similar it will generate the namespace file for you, including imports. I usually put all these statements in the package level man file, which you have `rgeospatialquality.R`. See https://github.com/ropensci/rgbif/blob/master/R/rgbif-package.r#L28-L41 for example
* It appears that you perhaps don't use `rgbif` in your package? If so, put it in `Suggests`, see
related problem during check:

```r
Namespace in Imports field not imported from: ‘rgbif’
  All declared Imports should be used.
```

If `rgbif` is moved to `Suggests`, make sure to use it conditionally in your examples, vignettes, etc. That is, e.g.

```r
if (requireNamespace("rgbif", quietly = TRUE)) {
  library("rgbif")
  ## your example code
} else {
  ## your example code if rgbif not available, or skip this else if rgbif required for example
}
```

## Continuous integration

Nice work having Travis already set up. I'd also suggest using Appveyor-CI, see `devtools::use_appveyor()`. I forked your package and ran on Appveyor, tehre are

I got a failing test on windows, see https://ci.appveyor.com/project/sckott/rgeospatialquality/build/1.0.1#L793, but i'm not sure if that will be fixed when your Imports are structued correctly or not.

## Docs

look good

## Tests

* All tests pass. nice. coverage is at about 65%, pretty good, but work on getting it up through time.
* I imagine `add_flags()` is a good candidate for an S3 class with methods for at least
data.frame and list. you have examples for data.frame's. Is there any use case where users
would have lists instead of data.frame's? If so, could have `add_flags.data.frame()` and
`add_flags.list()`
* I'd like to see more tests especially covering how the package should fail, so e.g. using `testthat::expect_error` or similar to indicate how the functions should fail, e.g., what error messages the functions should give, whether they give back no content or they do and what it should be when they fail. Some example behavior I'm not quite sure I understand:

What does it mean to pass no inputs and get back data?

```r
parse_record()
#> $hasCoordinates
#> [1] FALSE
#>
#> $hasScientificName
#> [1] FALSE
#>
#> $hasCountry
#> [1] FALSE
```

## Examples

The examples are a bit sparse. You should add more, covering the minimum set of
different usage scenarios with the various parameter inputs.

## Specific comments about the scripts and functions:

- curl options: my personal preference is that all functions that make http requests allow users
to pass in curl options (e.g,. verbose, progress, timeout, etc.). I'd recommend adding this
ability to the appropriate fxns. I usually do this with `...`, then users can just use `httr`
curl options as if they were calling `GET`/`POST` directly, but if `...` isn't possible, then
use a named parameter.
- you use `NA` for default parameters in your functions. I think it's better practice to
use `NULL` instead. though maybe you had a good reason to use `NA`? You can easily then remove
all `NULL` params input before passing into `GET` or `POST` query by using a utility function
`plyr::compact`, all it is is `function (x) Filter(Negate(is.null), x)`, which you can just
put in your package
- `format_gq()`
    - I would provide a `quiet` or `verbose` parameter so that the messages
can be silenced in case the user wants to do that.
    - you say the `indf` parameter is required, however you have a default of `NA` for that
    parameter. Just use `indf` in the function definition, so users that don't pass that
    parameter get an error that it's missing
    - You should check user inputs to the `source` parameter, e.g., `format_gq(d, source="asdf")`
    doesn't fail well. Use e.g., `match.arg()` or something similar
    - I think in addition to supplying `source` parameter you could auto-detect
    the data source (gbif, vertnet, etc.) at least with those data sources that include
    class info or other info that uniquely identifies where it came from. And if auto-detection
    of source fails you could error with message
- `parse_record()`
    - at the top of the function you have `if (!missing(record))`, but you have the default for
    that param set to `NA`, so how could it be missing?  See also above comment about switching
    to `NULL` as defaults, and if you change to `NULL`, the helper function `gq_parse_params()`
    I think is just replaced by the `compact()` function above, which just removes any that are
    `NULL`
    - Users may pass in a list to the first parameter as well as inputs to the others. You
    could check internally whether both are passed, and stop with informative message if they are
- `add_flags()`
    - same comment above for `indf` parameter, if required, don't give a default
- `gq_parse` has code for parsing error messages. would be nice to see tests showing
what is expected in different scenarios
- I wonder if you could provide some flexibility to users in the required fields input
to `add_flags()`, e.g., you could check for `decimalLatitude` ignoring case, and even accept something
like `latitude` or `lat`. Not sure if there's limitations here due to the API though.

## Code of conduct:

- do add a code of conduct, see `devtools::use_code_of_conduct()`

## est. hours

~5
