### 2016-03-28 #28 @jotegui
Thanks for the comments, @sckott. Really useful! I will start working on them and will update this thread when finished. Shouldn't take too long.
