### 2016-04-01 #28 @jotegui
Hi @sckott. I have a new version ready (v0.3.0) addressing your suggestions. I'll go over them sequentially. Please let me know if I need to do anything else for the review.

Thanks!

---

> - `Title` field: you have `R Wrapper for the Geospatial Data Quality REST API`. Just change to `Wrapper for the Geospatial Data Quality REST API`, as they sometimes complain about the use of `R` in the title and description, since they say it's obvious it's for R since its on CRAN.

done

> - `Description` field: you have `This package provides native wrappers for the functions available via de spatial quality REST API. More information on the API can be found here: http://bit.ly/bioinformatics_btw057.`. Change `de` to `the`

done

> I suggest removing
```
r_github_packages:
  - jimhester/covr
```
and put `covr` in `Suggests` in `DESCRIPTION` file

done

> Any reason you went with GPL? We usually default to MIT unless good reason to do something else. Just curious :)
also, docs https://cran.r-project.org/doc/manuals/r-release/R-exts.html#Licensing i think suggest that what you have (`file LICENSE`) isn't quite valid

Not really. It's just the license I have been using so far for my other projects. Happy to change it to MIT, for intra-ropensci consistency's sake :) I have adapted the `DESCRIPTION` file accordingly.

> * You have 3 packges listed as imports, but you don't have roxygen import statements in your package.

fixed. As you suggest, I have added `@importFrom` statements in the `rgeospatialquality.R` file

> * It appears that you perhaps don't use `rgbif` in your package? If so, put it in `Suggests`, see related problem during check:

done. I followed your suggestion and use `requireNamespace` each time `rgbif` (or any other data source package) is used. Now (as you will see below and in the package's `README`, both Travis and Appveyor pass.

> I'd also suggest using Appveyor-CI

done

> I got a failing test on windows, see https://ci.appveyor.com/project/sckott/rgeospatialquality/build/1.0.1#L793, but i'm not sure if that will be fixed when your Imports are structued correctly or not.

fixed. I have successfully integrated Appveyor now

> * All tests pass. nice. coverage is at about 65%, pretty good, but work on getting it up through time.

Current coverage is 93%

> * I imagine `add_flags()` is a good candidate for an S3 class with methods for at least data.frame and list. you have examples for data.frame's. Is there any use case where users would have lists instead of data.frame's? If so, could have `add_flags.data.frame()` and `add_flags.list()`

It is probably because of my (almost total) lack of understanding of how S3 classes work, but I can't see how can this improve usability of the function. In principle, `add_flags` is meant to work with datasets of more than 1 record, and therefore, the structure that would fit best here is the `data.frame`. Aren't `list`s designed just for the equivalent of a single record? If so, `parse_record` would be the best function to use here... But, as I said, it might be my lack of understanding, so if you can help me seeing it better, I can definitely work on turning this an S3 class :)

> * I'd like to see more tests especially covering how the package should fail (...)

Due to the flexible nature of the underlying API, it is actually difficult to make it fail, and there is almost none error handling currently implemented. However, I have added some warnings and improved docs to reflect this.

> What does it mean to pass no inputs and get back data?
```r
parse_record()
#> $hasCoordinates
#> [1] FALSE
#>
#> $hasScientificName
#> [1] FALSE
#>
#> $hasCountry
#> [1] FALSE
```

I have added warning messages and docs explaining this more generally. In this particular case, we wanted the API to be flexible and, even thought this is a feasible call, it is totally useless.

> The examples are a bit sparse

done

> - curl options: my personal preference is that all functions that make http requests allow users to pass in curl options (e.g,. verbose, progress, timeout, etc.). I'd recommend adding this ability to the appropriate fxns. I usually do this with `...`, then users can just use `httr` curl options as if they were calling `GET`/`POST` directly, but if `...` isn't possible, then use a named parameter.

done, added `...` to every `httr` calling function

> - you use `NA` for default parameters in your functions. I think it's better practice to use `NULL` instead. though maybe you had a good reason to use `NA`?

done. Not any particular reason to use `NA`

> `format_gq`

>     - I would provide a `quiet` or `verbose` parameter so that the messages can be silenced in case the user wants to do that.

done

>     - you say the `indf` parameter is required, however you have a default of `NA` for that parameter. Just use `indf` in the function definition, so users that don't pass that parameter get an error that it's missing

done

>     - You should check user inputs to the `source` parameter, e.g., `format_gq(d, source=\"asdf\")` doesn't fail well. Use e.g., `match.arg()` or something similar

done

>     - I think in addition to supplying `source` parameter you could auto-detect the data source (gbif, vertnet, etc.) at least with those data sources that include class info or other info that uniquely identifies where it came from. And if auto-detection of source fails you could error with message

I also thought about it, but the problem is that (among the tested packages) only `rgbif` shows any distinctive info when calling `class(obj)`. The other two (`rvertnet` and `rinat`) only return `data.frame` or `list`. And I don't really feel comfortable relying on the `data.frame`s' headers to figure it out, since that is likely to change. If it is a really helpful addition, I can do that, but it will probably only work for `rgbif`

> - `parse_record()`

>     - at the top of the function you have `if (!missing(record))`, but you have the default for that param set to `NA`, so how could it be missing?

fixed

> See also above comment about switching to `NULL` as defaults

done

>     - Users may pass in a list to the first parameter as well as inputs to the others. You could check internally whether both are passed, and stop with informative message if they are

fixed

> - `add_flags()`

>     - same comment above for `indf` parameter, if required, don't give a default

done

> - `gq_parse` has code for parsing error messages. would be nice to see tests showing  what is expected in different scenarios

See comment above re: making the API fail. It is difficult to create scenarios where the API might throw an error message...

> - I wonder if you could provide some flexibility to users in the required fields input to `add_flags()`, e.g., you could check for `decimalLatitude` ignoring case, and even accept something like `latitude` or `lat`. Not sure if there's limitations here due to the API though.

done. Added a new parameter to `add_flags()`, called `guess_fields`, that does this

> - do add a code of conduct, see `devtools::use_code_of_conduct()`

done
