### 2016-04-02 #28 @jotegui
Awesome, many thanks!

--
Javier Otegui
http://www.jotegui.com
El 02/04/2016 21:13, "Scott Chamberlain" <notifications@github.com>
escribió:

> thanks, i'll have a look soon and get back to you here
>
