### 2016-04-07 #28 @sckott
Thanks for all the changes. Just commenting on those that require responses:

> Happy to change it to MIT, for intra-ropensci consistency's sake :) I have adapted the DESCRIPTION file accordingly.

thanks!

>  It is probably because of my (almost total) lack of understanding of how S3 classes work, but I can't see how can this improve usability of the function. In principle, add_flags is meant to work with datasets of more than 1 record, and therefore, the structure that would fit best here is the data.frame. Aren't lists designed just for the equivalent of a single record? If so, parse_record would be the best function to use here... But, as I said, it might be my lack of understanding, so if you can help me seeing it better, I can definitely work on turning this an S3 class :)

lists can be of any length, and nested as well. I don't think you have to support lists, but if you only want to support data.frame's then I'd make an S3 class so that the function fails nicely for users that try to pass other classes into it. e.g.,

```r
foo <- function(x) {
    UseMethod("foo")
}

foo.data.frame <- function(x) {
    # do stuff
}

foo.default <- function(x) {
    stop("foo does not support input of class ", class(x), call. = FALSE)
}
```

> I also thought about it, but the problem is that (among the tested packages) only rgbif shows any distinctive info when calling class(obj). ...

Okay, makes sense
