### 2016-04-07 #28 @sckott
@jotegui after resolution of the above comment about S3 - we're ready to move it to ropensci
