### 2016-04-08 #28 @jotegui
Hi @sckott

Thanks for your answer, and for the hint on S3 classes and methods. I have now updated the code and implemented both `add_flags` and `format_gq` as S3 classes, with methods only for `data.frames`. I thought it was better to release this early and, later on, I can always add new methods for different object types, like `list`s, right?

Also, I have added a bunch of tests for this new setup and fixed some weird dependency issues with travis. I had to add the following lines to `.travis.yml` to make sure it finds the packages under `Suggests`:

```yaml
repos:
  CRAN: https://cloud.r-project.org
  ropensci: http://packages.ropensci.org
r_github_packages:
  - ropensci/rgbif
  - ropensci/rvertnet
  - ropensci/rinat
```

Let me know if I need to do something else.

Cheers!
