### 2016-04-08 #28 @sckott
> I have now updated the code and implemented both add_flags and format_gq as S3 classes, with methods only for data.frames

great, thanks

> I thought it was better to release this early and, later on, I can always add new methods for different object types, like lists, right?

Yeah, easy to add functionality later

Keep in mind with `Remotes` this https://github.com/hadley/devtools/blob/master/vignettes/dependencies.Rmd#cran-submission
