### 2016-04-08 #28 @sckott
I think we're all set - I invited you to our organization `ropenscilabs` - we start newer pkgs there now, then move to `ropensci` when pkgs are more stable - After you accept the invitation, you should be able to transfer the repo, let me know if that doesn't work
